requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
astroid==3.3.5
attrs==24.2.0
autopep8==2.3.1
backports.tarfile==1.2.0
black==24.8.0
bleach==6.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
cssselect==1.2.0
decorator==5.1.1
dicttoxml==1.7.16
dill==0.3.9
docutils==0.21.2
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
geoip2==4.8.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ip2geotools==0.1.6
IP2Location==8.10.4
ipinfo==5.1.1
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.4.1
lazy-object-proxy==1.10.0
lxml==5.3.0
markdown-it-py==3.0.0
maxminddb==2.6.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
mypy-extensions==1.0.0
nh3==0.2.18
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
pip-review==1.3.0
pkginfo==1.10.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.0
pycodestyle==2.12.1
pycparser==2.22
Pygments==2.18.0
pylint==3.3.1
pyparsing==3.1.4
pyquery==2.0.1
PySocks==1.7.1
pytest==8.3.3
python-dotenv==1.0.1
ratelim==0.1.6
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.2
SecretStorage==3.3.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomlkit==0.13.2
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
twine==5.1.1
typed-ast==1.5.5
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.3
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.14.0
zipp==3.20.2
